add chart header
