+++ tailwind.config.js
@@ -42,36 +42,6 @@ module.exports = {
           inverted: colors.white,
         },
       },
-      // dark mode
-      'dark-tremor': {
-        brand: {
-          faint: '#0B1229',
-          muted: colors.blue[950],
-          subtle: colors.blue[800],
-          DEFAULT: colors.blue[500],
-          emphasis: colors.blue[400],
-          inverted: colors.blue[950],
-        },
-        background: {
-          muted: '#131A2B',
-          subtle: colors.gray[800],
-          DEFAULT: colors.gray[900],
-          emphasis: colors.gray[300],
-        },
-        border: {
-          DEFAULT: colors.gray[800],
-        },
-        ring: {
-          DEFAULT: colors.gray[800],
-        },
-        content: {
-          subtle: colors.gray[600],
-          DEFAULT: colors.gray[500],
-          emphasis: colors.gray[200],
-          strong: colors.gray[50],
-          inverted: colors.gray[950],
-        },
-      },
       boxShadow: {
         // light
         'tremor-input': '0 1px 2px 0 rgb(0 0 0 / 0.05)',
